requirements.txt
amqp==2.5.2
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-six==1.0.4
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.17.0
pytz==2021.1
simple-settings==1.0.0
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
